Crate remote_controller

Remote control

This is a very hack web interface for hamilton. It’s really not great. Look at the rest of the project not here!

Structs

Action

ActionIdWrapper

ActionList

AreaSize

CanvasTouch

ControllerState

GamepadCommand

Functions

start_remote_controller_server

start_remote_controller_server_with_map